refactor: テーマ入力画面のリファクタリング #175
