+++ front/src/app/[uuid]/input-theme/page.tsx
@@ -1,8 +1,8 @@
-import InputThemePresentation from "@/app/presentation/InputTheme/InputThemePresentation";
+import InputThemeContainer from "@/app/[uuid]/input-theme/InputThemeContainer";
 import { getIdeaSessionInProgress } from "@/lib/idea-sessions";
 
 export default async function InputTheme() {
   const ideaSession = await getIdeaSessionInProgress();
 
-  return <InputThemePresentation ideaSession={ideaSession} />;
+  return <InputThemeContainer ideaSession={ideaSession} />;
 }
