+++ front/src/components/elements/AlertModal/AlertModal.tsx
@@ -1,0 +1,35 @@
+"use client";
+
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+} from "@/components/ui/alert-dialog";
+
+export default function AlertModal({
+  isOpen,
+  onClick,
+  children,
+}: {
+  isOpen: boolean;
+  onClick: () => void;
+  children: React.ReactNode;
+}) {
+  return (
+    <AlertDialog open={isOpen} aria-labelledby="responsive-dialog-title">
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogDescription className="text-left">
+            {children}
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter className="flex-col sm:flex-row gap-4 sm:gap-0">
+          <AlertDialogAction onClick={onClick}>OK</AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
