+++ front/src/features/input-theme/components/Form/Form.module.scss
@@ -1,42 +1,3 @@
-.wrapper {
-  min-height: calc(100vh - var(--header-height));
-  width: 100vw;
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
-  height: calc(100% - var(--back-button-length) - var(--page-padding));
-  width: calc(100% - var(--page-padding) * 2);
-  padding-bottom: calc(var(--back-button-length) + var(--page-padding) + 80px);
-
-  @media screen and (min-width: 1024px) {
-    height: calc(100% - var(--back-button-length) - var(--page-padding-lg));
-    width: calc(100% - var(--page-padding-lg) * 2);
-    padding-bottom: calc(
-      var(--back-button-length) + var(--page-padding-lg) + 160px
-    );
-  }
-}
-
-.content {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  @media screen and (min-width: 768px) {
-    width: 50%;
-  }
-}
-
 /* フォーム */
 .form {
   display: flex;
@@ -121,24 +82,4 @@
   @media screen and (min-width: 1024px) {
     font-size: 16px;
   }
-}
-
-/* 戻るボタン */
-.back {
-  position: absolute;
-  bottom: var(--page-padding);
-  left: var(--page-padding);
-
-  @media screen and (min-width: 1024px) {
-    bottom: var(--page-padding-lg);
-    left: var(--page-padding-lg);
-  }
-}
-
-.arrow {
-  position: absolute;
-  top: 0;
-  left: 0;
-  height: 100%;
-  margin-left: 0.25rem;
 }
