+++ front/src/features/input-theme/components/index.ts
@@ -1,0 +1,3 @@
+import Form from "@/features/input-theme/components/Form/Form";
+
+export { Form };
